refactor: type 定義をディレクトリ内に移動
